refactor(tests): co-locate orphan src/lib root tests with their targets
Phase C of directory-structure-review. Three test files sat at src/lib/ root
while their subjects live in subdirs; move each next to its target via the
phase-config codemod (relative imports auto-rewritten):
- callback-url-basepath.test.ts -> src/lib/auth/session/
- vault-unlock-error.test.ts -> src/lib/vault/
- vault-context-loading-timeout.test.tsx -> src/lib/vault/

The live vi.mock specifiers were converted to @/lib aliases in the preceding
prep commit so the mocks keep intercepting from the new location.
