+++ src/lib/auth/session/callback-url-basepath.test.ts
@@ -9,7 +9,7 @@ vi.mock("@/lib/url-helpers", () => ({
   BASE_PATH: "/passwd-sso",
 }));
 
-import { callbackUrlToHref } from "./auth/session/callback-url";
+import { callbackUrlToHref } from "./callback-url";
 
 describe("callbackUrlToHref (BASE_PATH=/passwd-sso)", () => {
   it("strips basePath and locale prefix", () => {
